Stronger Types

+++ src/attending-server/base-attending-server.ts
@@ -7,7 +7,6 @@ import {
     BaseMessageOptions,
     TextChannel,
     User,
-    VoiceChannel,
     VoiceState,
     ChannelType,
     VoiceBasedChannel
@@ -23,7 +22,13 @@ import { FirebaseServerBackupExtension } from './firebase-backup.js';
 import { CalendarExtensionState } from '../extensions/session-calendar/calendar-states.js';
 import { QueueBackup } from '../models/backups.js';
 import { blue, cyan, green, magenta, red, yellow } from '../utils/command-line-colors.js';
-import { convertMsToTime } from '../utils/util-functions.js';
+import {
+    convertMsToTime,
+    isCategoryChannel,
+    isQueueTextChannel,
+    isTextChannel,
+    isVoiceChannel
+} from '../utils/util-functions.js';
 import {
     CategoryChannelId,
     GuildMemberId,
@@ -61,8 +66,6 @@ class AttendingServerV2 {
     private queueChannelsCache: QueueChannel[] = [];
     /** unique active helpers, key is member.id */
     private _activeHelpers: Collection<GuildMemberId, Helper> = new Collection();
-    /** enables helper vc status in queue embeds */
-    private readonly useExperimentalVCStatusRerender = true;
 
     protected constructor(
         readonly user: User,
@@ -100,8 +103,8 @@ class AttendingServerV2 {
             return false;
         }
         await Promise.all([
-            this.serverExtensions.map(extension => extension.onServerRequestBackup(this)),
-            this._queues.map(queue => queue.setSeriousMode(enableSeriousMode))
+            this._queues.map(queue => queue.setSeriousMode(enableSeriousMode)),
+            this.serverExtensions.map(extension => extension.onServerRequestBackup(this))
         ]);
         return true;
     }
@@ -155,6 +158,8 @@ class AttendingServerV2 {
                   new FirebaseServerBackupExtension(guild),
                   CalendarExtensionState.load(guild)
               ]);
+        // create instance with extensions
+        const server = new AttendingServerV2(user, guild, serverExtensions);
         // Retrieve backup from all sources. Take the first one that's not undefined
         // TODO: Change behavior here depending on backup strategy
         const externalBackup = environment.disableExtensions
@@ -164,16 +169,17 @@ class AttendingServerV2 {
                       extension.loadExternalServerData(guild.id)
                   )
               );
+        // now load the data from backup
         const externalServerData = externalBackup?.find(backup => backup !== undefined);
         if (externalServerData !== undefined) {
             console.log(cyan(`Found external backup for ${guild.name}. Restoring.`));
-        }
-        const server = new AttendingServerV2(user, guild, serverExtensions);
-        server._afterSessionMessage = externalServerData?.afterSessionMessage ?? '';
-        if (externalServerData?.loggingChannelId !== undefined) {
-            server._loggingChannel = server.guild.channels.cache.get(
-                externalServerData?.loggingChannelId
-            ) as TextChannel;
+            const loggingChannelFromBackup = server.guild.channels.cache.get(
+                externalServerData.loggingChannelId
+            );
+            if (isTextChannel(loggingChannelFromBackup)) {
+                server._loggingChannel = loggingChannelFromBackup;
+            }
+            server._afterSessionMessage = externalServerData.afterSessionMessage;
         }
         // This call must block everything else for handling empty servers
         await server.createHierarchyRoles();
@@ -208,6 +214,11 @@ class AttendingServerV2 {
         member: GuildMember,
         newVoiceState: WithRequired<VoiceState, 'channel'>
     ): Promise<void> {
+        // temporary solution, stage channel is not currently supported
+        if (!isVoiceChannel(newVoiceState.channel)) {
+            return;
+        }
+        const vc = newVoiceState.channel;
         const memberIsStudent = this._activeHelpers.some(helper =>
             helper.helpedMembers.some(
                 helpedMember => helpedMember.member.id === member.id
@@ -225,15 +236,9 @@ class AttendingServerV2 {
             );
             await Promise.all([
                 ...this.serverExtensions.map(extension =>
-                    extension.onStudentJoinVC(
-                        this,
-                        member,
-                        // already checked
-                        newVoiceState.channel as VoiceChannel
-                    )
+                    extension.onStudentJoinVC(this, member, vc)
                 ),
-                ...(this.useExperimentalVCStatusRerender &&
-                    queuesToRerender.map(queue => queue.triggerRender()))
+                ...queuesToRerender.map(queue => queue.triggerRender())
             ]);
         }
         if (memberIsHelper) {
@@ -280,8 +285,7 @@ class AttendingServerV2 {
                 ),
                 this.afterSessionMessage !== '' &&
                     member.send(SimpleEmbed(this.afterSessionMessage)),
-                ...(this.useExperimentalVCStatusRerender &&
-                    queuesToRerender.map(queue => queue.triggerRender()))
+                ...queuesToRerender.map(queue => queue.triggerRender())
             ]);
         }
         if (memberIsHelper) {
@@ -294,6 +298,7 @@ class AttendingServerV2 {
             );
         }
     }
+
     /**
      * Gets all the queue channels on the server. SLOW
      * if nothing is found, returns empty array
@@ -306,26 +311,22 @@ class AttendingServerV2 {
         }
         const allChannels = await this.guild.channels.fetch();
         // cache again on a fresh request
-        this.queueChannelsCache = allChannels
-            .filter(ch => ch !== null && ch.type === ChannelType.GuildCategory)
-            // ch has type 'AnyChannel', have to cast, type already checked
-            .map(channel => channel as CategoryChannel)
-            .map(category => [
-                category.children.cache.find(
-                    child =>
-                        child.name === 'queue' && child.type === ChannelType.GuildText
-                ),
-                category.name,
-                category.id
-            ])
-            .filter(([textChannel]) => textChannel !== undefined)
-            .map(([ch, name, parentId]) => {
-                return {
-                    channelObj: ch,
-                    queueName: name,
-                    parentCategoryId: parentId
-                } as QueueChannel;
+        this.queueChannelsCache = [];
+        for (const categoryChannel of allChannels.values()) {
+            if (!isCategoryChannel(categoryChannel)) {
+                continue;
+            }
+            const queueTextChannel: Optional<TextChannel> =
+                categoryChannel.children.cache.find(isQueueTextChannel);
+            if (!queueTextChannel) {
+                continue;
+            }
+            this.queueChannelsCache.push({
+                channelObj: queueTextChannel,
+                queueName: categoryChannel.name,
+                parentCategoryId: categoryChannel.id
             });
+        }
         const duplicateQueues = this.queueChannelsCache
             .map(queue => queue.queueName)
             .filter((item, index, arr) => arr.indexOf(item) !== index);
@@ -370,6 +371,7 @@ class AttendingServerV2 {
             this.createClassRoles()
         ]);
         this._queues.set(parentCategory.id, helpQueue);
+        await this.getQueueChannels(false);
     }
 
     /**
@@ -383,36 +385,43 @@ class AttendingServerV2 {
         if (queue === undefined) {
             throw ExpectedServerErrors.queueDoesNotExist;
         }
-        const parentCategory = (await (await this.guild.channels.fetch())
-            .find(ch => ch !== null && ch.id === queueCategoryID)
-            ?.fetch()) as CategoryChannel;
-        // delete child channels first
-        await Promise.all(
-            parentCategory?.children.cache
-                .map(child => child.delete())
-                .filter(promise => promise !== undefined)
+        // delete queue data model no matter if the category was deleted by the user
+        // now only the queue variable holds the queue channel
+        this._queues.delete(queueCategoryID);
+        const allChannels = await this.guild.channels.fetch();
+        const parentCategory = allChannels.find(
+            (channel): channel is CategoryChannel =>
+                isCategoryChannel(channel) && channel.id === queueCategoryID
         );
+        if (!parentCategory) {
+            // this shouldn't happen bc input type restriction on the command
+            // but we need to pass ts check
+            throw ExpectedServerErrors.queueDoesNotExist;
+        }
+        // delete child channels first
+        await Promise.all(parentCategory.children.cache.map(child => child.delete()));
         // now delete category, role, and let queue call onQueueDelete
         await Promise.all([
-            parentCategory?.delete(),
-            (await this.guild.roles.fetch())
+            parentCategory.delete(),
+            this.guild.roles.cache
                 .find(role => role.name === parentCategory.name)
                 ?.delete(),
             queue.gracefulDelete()
         ]);
-        // finally delete queue data model and refresh cache
-        this._queues.delete(queueCategoryID);
         await this.getQueueChannels(false);
     }
 
     /**
      * Attempt to enqueue a student
      * @param studentMember student member to enqueue
-     * @param queue target queue
-     * @throws QueueError: if @param queue rejects
+     * @param queueChannel target queue
+     * @throws QueueError: if queue rejects
      */
-    async enqueueStudent(studentMember: GuildMember, queue: QueueChannel): Promise<void> {
-        await this._queues.get(queue.parentCategoryId)?.enqueue(studentMember);
+    async enqueueStudent(
+        studentMember: GuildMember,
+        queueChannel: QueueChannel
+    ): Promise<void> {
+        await this._queues.get(queueChannel.parentCategoryId)?.enqueue(studentMember);
         await Promise.all(
             this.serverExtensions.map(extension => extension.onServerRequestBackup(this))
         );
@@ -454,7 +463,8 @@ class AttendingServerV2 {
                     : curr //  both prev.first and curr.first will not be undefined
         );
         const student = await queueToDequeue.dequeueWithHelper(helperMember);
-        await this.sendInvite(helperObject, student, helperVoiceChannel);
+        helperObject.helpedMembers.push(student);
+        await this.sendInvite(student, helperVoiceChannel);
         return student;
     }
 
@@ -506,7 +516,8 @@ class AttendingServerV2 {
         } else {
             throw ExpectedServerErrors.badDequeueArguments;
         }
-        await this.sendInvite(helperObject, student, helperVoiceChannel);
+        helperObject.helpedMembers.push(student);
+        await this.sendInvite(student, helperVoiceChannel);
         return student;
     }
 
@@ -752,16 +763,14 @@ class AttendingServerV2 {
      */
     async updateCommandHelpChannels(): Promise<void> {
         const allChannels = await this.guild.channels.fetch();
-        const existingHelpCategory = allChannels
-            .filter(
-                channel =>
-                    channel !== null &&
-                    channel.type === ChannelType.GuildCategory &&
-                    channel.name === 'Bot Commands Help'
-            )
-            .map(channel => channel as CategoryChannel);
+        const existingHelpCategory = allChannels.find(
+            (channel): channel is CategoryChannel =>
+                channel !== null &&
+                channel.type === ChannelType.GuildCategory &&
+                channel.name === 'Bot Commands Help'
+        );
         // If no help category is found, initialize
-        if (existingHelpCategory.length === 0) {
+        if (!existingHelpCategory) {
             console.log(
                 cyan(`Found no help channels in ${this.guild.name}. Creating new ones.`)
             );
@@ -769,7 +778,6 @@ class AttendingServerV2 {
                 name: 'Bot Commands Help',
                 type: ChannelType.GuildCategory
             });
-            existingHelpCategory.push(helpCategory);
             // Change the config object and add more functions here if needed
             await Promise.all(
                 commandChConfigs.map(async roleConfig => {
@@ -797,14 +805,18 @@ class AttendingServerV2 {
                     );
                 })
             );
+            await this.sendCommandHelpChannelMessages(helpCategory, commandChConfigs);
         } else {
             console.log(
-                yellow(
-                    `Found existing help channels in ${this.guild.name}, updating command help files`
-                )
+                `Found existing help channels in ${yellow(
+                    this.guild.name
+                )}, updating command help files`
+            );
+            await this.sendCommandHelpChannelMessages(
+                existingHelpCategory,
+                commandChConfigs
             );
         }
-        await this.sendCommandHelpChannelMessages(existingHelpCategory, commandChConfigs);
         console.log(magenta(`✓ Updated help channels on ${this.guild.name} ✓`));
     }
 
@@ -842,11 +854,9 @@ class AttendingServerV2 {
      * @param helperVoiceChannel
      */
     private async sendInvite(
-        helperObject: Helper,
         student: Readonly<Helpee>,
         helperVoiceChannel: VoiceBasedChannel
     ) {
-        helperObject.helpedMembers.push(student);
         const [invite] = await Promise.all([
             helperVoiceChannel.createInvite({
                 maxAge: 15 * 60,
@@ -979,36 +989,30 @@ class AttendingServerV2 {
         const existingRoles = new Set(this.guild.roles.cache.map(role => role.name));
         const queueNames = (await this.getQueueChannels(false)).map(ch => ch.queueName);
         await Promise.all(
-            queueNames
-                .filter(queue => !existingRoles.has(queue))
-                .map(roleToCreate =>
+            queueNames.map(roleToCreate => {
+                !existingRoles.has(roleToCreate) &&
                     this.guild.roles.create({
                         name: roleToCreate,
                         position: 1
-                    })
-                )
+                    });
+            })
         );
     }
 
     /**
      * Overwrites the existing command help channel and send new help messages
-     * @param helpCategories the category named 'Bot Commands Help'
+     * @param helpCategory the category named 'Bot Commands Help'
      * @param messageContents array of embeds to send to each help channel
      */
     private async sendCommandHelpChannelMessages(
-        helpCategories: CategoryChannel[],
+        helpCategory: CategoryChannel,
         messageContents: Array<{
             channelName: string;
             file: HelpMessage[];
             visibility: string[];
         }>
     ): Promise<void> {
-        const allHelpChannels = helpCategories.flatMap(
-            category =>
-                [...category.children.cache.values()].filter(
-                    ch => ch.type === ChannelType.GuildText
-                ) as TextChannel[]
-        );
+        const allHelpChannels = helpCategory.children.cache.filter(isTextChannel);
         await Promise.all(
             allHelpChannels.map(
                 async ch =>
@@ -1025,6 +1029,9 @@ class AttendingServerV2 {
                     ?.file?.filter(helpMessage => helpMessage.useInHelpChannel)
                     .map(helpMessage => channel.send(helpMessage.message))
             )
+        );
+        console.log(
+            `Successfully updated help messages in ${yellow(helpCategory.name)}!`
         );
     }
 }
